+++ bld/f77/wfc/h/gstring.h
@@ -2,6 +2,7 @@
 *
 *                            Open Watcom Project
 *
+* Copyright (c) 2002-2021 The Open Watcom Contributors. All Rights Reserved.
 *    Portions Copyright (c) 1983-2002 Sybase, Inc. All Rights Reserved.
 *
 *  ========================================================================
@@ -29,8 +30,8 @@
 ****************************************************************************/
 
 
-extern sym_id  GStartCat( uint num_args, uint size );
-extern sym_id  GTempString( uint size );
+extern sym_id  GStartCat( uint num_args, size_t size );
+extern sym_id  GTempString( size_t size );
 extern void    GStopCat( uint num_args, sym_id result );
 extern void    GCatArg( itnode *itptr );
 extern void    AsgnChar( void );
